transaction/history: cleanup godoc
GitOrigin-RevId: eef27f4c93c93f25e2378e20572dac72456db709

+++ pkg/transaction/history/model/util.go
@@ -107,6 +107,6 @@ func (m *Entry) GetOrderingKey() ([]byte, error) {
 func OrderKeyFromSequence(v KinVersion, seq uint32) []byte {
 	var b [5]byte
 	b[0] = byte(v)
-	binary.BigEndian.PutUint32(b[1:], uint32(seq))
+	binary.BigEndian.PutUint32(b[1:], seq)
 	return b[:]
 }
